extracted class type from flight class

+++ src/main/java/com/twair/BasePrice.java
@@ -5,12 +5,12 @@ import java.util.List;
 import java.util.Map;
 
 public class BasePrice {
-    public int calculate(Flight flight, ClassType classType, int numberOfSeats) {
+    public Double calculate(Flight flight, ClassType classType, int numberOfSeats) {
         return flight.getBasePrice(classType) * numberOfSeats;
     }
 
-    public Map<String, Integer> calculate(List<Flight> flights, ClassType classType, int numberOfSeats) {
-        Map<String, Integer> priceMap = new HashMap<>();
+    public Map<String, Double> calculate(List<Flight> flights, ClassType classType, int numberOfSeats) {
+        Map<String, Double> priceMap = new HashMap<>();
         for(Flight flight : flights) {
             priceMap.put(flight.getNumber(), calculate(flight, classType, numberOfSeats));
         }
